0.0.5
Added more adjustments to different screen sizes.

Known Errors: on some iPhones the top right title span is falling to the bottom of the card deck for some unknown reason.

Firebase also is for some reason not updating the main site.

However the code is updated and working as intended.